% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.R
\name{detail_view}
\alias{detail_view}
\alias{get_cat}
\alias{get_aut}
\alias{get_lim}
\title{List views}
\usage{
get_cat(id)

get_aut(id)

get_lim(id)
}
\arguments{
\item{id}{id of object to retrieve}
}
\description{
Get categories, authors or limericks of limericks available.
}
\examples{
\dontrun{
cat <- get_cat(id = 1) # get limericks in first category
aut <- get_aut(id = 1) # get limericks by authord with ID 1
lim <- get_lim(id = 1) # get details of limerick ID 1
}

}
\seealso{
\code{\link{list_view}}
}
