% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{fl_auth}
\alias{fl_auth}
\title{Authenticate}
\usage{
fl_auth(username, password, verbose = FALSE)
}
\arguments{
\item{username}{your username, see details.}

\item{password}{your password, see details.}

\item{verbose}{boolean if you want feedback on call.}
}
\description{
five-line authentication, retrieve token
}
\details{
\code{username} and \code{password} can be obtained by visiting
    \href{http://www.five-line.org}{five-line} and creating an account.
    Then simply use your username and password in this function.
    Tokens are valid for 30 minutes.
}
\examples{
\dontrun{
lim_auth(username = "my_username", password = "my_password")
}

}
\seealso{
\href{http://www.five-line.org}{five-line}
}
