% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_lim}
\alias{insert_lim}
\title{Insert limerick in RMarkdown}
\usage{
insert_lim(limerick, title = TRUE)
}
\arguments{
\item{limerick}{A limerick as returned by \code{\link{get_lim}}}

\item{title}{whether to include title}
}
\description{
Insert limerick in RMarkdown
}
\examples{
\dontrun{
lim <- get_lim(id = 1)
insert_lim(lim)
}

}
\seealso{
\code{\link{get_lim}}
}
