% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_view}
\alias{list_view}
\alias{list_cat}
\alias{list_aut}
\alias{list_lim}
\title{List views}
\usage{
list_cat()

list_aut()

list_lim()
}
\description{
List categories, authors or limericks of limericks available.
}
\examples{
\dontrun{
categories <- list_cat()
authors <- list_aut()
limericks <- list_lim()
}

}
